--- core/navigation/src/main/java/com/porring/navigation/RouteModel.kt
@@ -1,22 +1,0 @@
-package com.porring.navigation
-
-import kotlinx.serialization.Serializable
-
-sealed interface Route
-
-sealed interface MainMenuRoute: Route {
-    @Serializable
-    data object Home: MainMenuRoute
-
-    @Serializable
-    data object Search: MainMenuRoute
-
-    @Serializable
-    data object Camera: MainMenuRoute
-
-    @Serializable
-    data object Follower: MainMenuRoute
-
-    @Serializable
-    data object My: MainMenuRoute
-}
