+++ feature/home/src/main/java/com/kolown/home/navigation/HomeNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.kolown.home.HomeRoute
-import com.porring.navigation.MainMenuRoute
+import com.kolown.navigation.MainMenuRoute
 
 fun NavController.navigateHome(navOptions: NavOptions) {
     navigate(MainMenuRoute.Home, navOptions)
