+++ feature/my/src/main/java/com/kolown/my/navigation/MyNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.kolown.my.MyRoute
-import com.porring.navigation.MainMenuRoute
+import com.kolown.navigation.MainMenuRoute
 
 fun NavController.navigateMy(navOptions: NavOptions) {
     navigate(MainMenuRoute.My, navOptions)
