+++ feature/search/src/main/java/com/kolown/search/navigation/SearchNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.porring.navigation.MainMenuRoute
+import com.kolown.navigation.MainMenuRoute
 import com.kolown.search.SearchRoute
 
 fun NavController.navigateSearch(navOptions: NavOptions) {
