+++ settings.gradle.kts
@@ -25,15 +25,14 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 include(":app")
 include(":core:model")
 include(":core:data")
-include(":feature:main")
 include(":core:domain")
+include(":feature:main")
 include(":feature:home")
 include(":feature:follower")
 include(":feature:my")
 include(":feature:search")
 include(":feature:camera")
+include(":feature:login")
 include(":core:navigation")
 include(":app-test-camera")
 include(":core:common")
-
-include(":feature:login")
